Create new set of utils
