+++ src/utils/language.ts
@@ -1,7 +1,3 @@
-export function noop() {
-  // empty
-}
-
 export function isUndefined(value: any): value is undefined {
   return value === undefined;
 }
